small improvements

+++ setup.sh
@@ -4,15 +4,18 @@
 
 for file in ${DOT_FILES[@]}
 do
-    ln -s $HOME/dotfiles/$file $HOME/$file
+    if [ ! -f "$HOME/$file" ]; then
+        ln -s $HOME/dotfiles/$file $HOME/$file
+    fi
 done
 ln -s $HOME/dotfiles/.tmux $HOME/.tmux
 
 #setup for nvim
-if [ -d "$HOME/.config" ]; then
+if [ ! -d "$HOME/.config" ]; then
     mkdir $HOME/.config
 fi
 
+# install vim config
 ln -s $HOME/dotfiles/nvim $HOME/.config/nvim
 ln -s $HOME/dotfiles/nvim $HOME/.vim
 ln -s $HOME/dotfiles/nvim/init.vim $HOME/.vimrc
